requirements.txt
#Todo

django
django-environ
Django==5.1.5
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
psycopg2
psycopg